+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.6.7-rc1 - ????/??/??
 
+- [FEATURE] Refactor logging setup across multiple modules to be able to send logs to a syslog server and have multiple handlers at the same time
 - [BUGFIX] Fix issues with the Ingress controller regarding reverse proxy settings when using multiple paths per rule and a template by adjusting the indexing logic to be configurable via the new `KUBERNETES_REVERSE_PROXY_SUFFIX_START` setting (default is `1` to keep backward compatibility)
 
 ## v1.6.6 - 2025/11/24